+ Refix some warning
git-svn-id: svn://svn.gna.org/svn/freeciv-warclient/trunk@852 4efe6140-7f1d-0410-9bd2-be1ee28408da

+++ utility/netintf.c
@@ -206,7 +206,7 @@ static void dns_result_callback(const unsigned char *addr,
 
   freelog(LOG_DEBUG, "drc dns_result_callback data=%p "
           " qtype=%d hostname=\"%s\" addrlen=%lu",
-          data, qtype, hostname, addrlen);
+          data, qtype, hostname, (long unsigned) addrlen);
   
   assert(ctx != NULL);
   assert(ctx->callback != NULL);
